add some fixes.

+++ audiothek.py
@@ -1,9 +1,11 @@
 import argparse
+import json
 import os
 import re
-import requests
 import sys
-import json
+
+import requests
+
 
 def main(url: str, folder: str):
     match = re.search(r"/(\d+)/?$", url)
@@ -11,26 +13,37 @@ def main(url: str, folder: str):
         id = match.group(1)
         downloadEpisodes(id, folder)
 
+
 def downloadEpisodes(id: str, folder: str):
-    query = open(os.path.join(os.path.dirname(os.path.abspath(
-        __file__)), "graphql", "ProgramSetEpisodesQuery.graphql")).read()
+    query = open(
+        os.path.join(
+            os.path.dirname(os.path.abspath(__file__)),
+            "graphql",
+            "ProgramSetEpisodesQuery.graphql",
+        )
+    ).read()
 
     # change query if url is "sammlung"
     if "sammlung" in url.lower():
-        query = open(os.path.join(os.path.dirname(os.path.abspath(
-            __file__)), "graphql", "editorialCollection.graphql")).read()
+        query = open(
+            os.path.join(
+                os.path.dirname(os.path.abspath(__file__)),
+                "graphql",
+                "editorialCollection.graphql",
+            )
+        ).read()
 
-    variables = {
-        "id": id
-    }
-    response = requests.get('https://api.ardaudiothek.de/graphql', params={
-        'query': query,
-        'variables': json.dumps(variables)
-    })
+    variables = {"id": id}
+    response = requests.get(
+        "https://api.ardaudiothek.de/graphql",
+        params={"query": query, "variables": json.dumps(variables)},
+    )
     response_json = response.json()
 
-    nodes = response_json.get("data").get(
-        "result").get("items").get("nodes")
+    results = response_json.get("data", {}).get("result", {})
+    nodes = []
+    if results:
+        nodes = results.get("items", {}).get("nodes", {})
 
     for index, node in enumerate(nodes):
         number = node["id"]
@@ -49,8 +62,14 @@ def downloadEpisodes(id: str, folder: str):
 
         # get image information
         image_url = node.get("image").get("url")
-        image_url = image_url.replace("{width}", "500")
-        mp3_url = node.get("audios")[0].get("downloadUrl")
+        image_url = image_url.replace("{width}", "2000")
+        image_url_x1 = node.get("image").get("url1X1")
+        image_url_x1 = image_url_x1.replace("{width}", "2000")
+        if not node.get("audios"):
+            continue
+        mp3_url = node.get("audios")[0].get("downloadUrl") or node.get("audios")[0].get(
+            "url"
+        )
         programset_id = node.get("programSet").get("id")
 
         program_path: str = os.path.join(folder, programset_id)
@@ -62,38 +81,73 @@ def downloadEpisodes(id: str, folder: str):
             except FileExistsError:
                 pass
             except Exception as e:
-                print("[Error] Couldn't create output directory!",
-                      file=sys.stderr)
+                print("[Error] Couldn't create output directory!", file=sys.stderr)
                 print(e, file=sys.stderr)
                 return
 
-            # write image
-            image_file_path = os.path.join(program_path, filename + '.jpg')
+            # write images
+            image_file_path = os.path.join(program_path, filename + ".jpg")
+            image_file_x1_path = os.path.join(program_path, filename + "_x1.jpg")
 
-            if os.path.exists(image_file_path) == False:
+            if not os.path.exists(image_file_path) or not os.path.exists(
+                image_file_x1_path
+            ):
                 response_image = requests.get(image_url)
-                with open(image_file_path, 'wb') as f:
+                with open(image_file_path, "wb") as f:
+                    f.write(response_image.content)
+
+            if not os.path.exists(image_file_x1_path):
+                response_image = requests.get(image_url_x1)
+                with open(image_file_x1_path, "wb") as f:
                     f.write(response_image.content)
 
             # write mp3
-            mp3_file_path = os.path.join(program_path, filename + '.mp3')
+            mp3_file_path = os.path.join(program_path, filename + ".mp3")
 
-            print("Download: " + str(index + 1) + " of " +
-                  str(len(nodes)) + " -> " + mp3_file_path)
-            if os.path.exists(mp3_file_path) == False:
+            print(
+                "Download: "
+                + str(index + 1)
+                + " of "
+                + str(len(nodes))
+                + " -> "
+                + mp3_file_path
+            )
+            if os.path.exists(mp3_file_path) == False and mp3_url:
                 response_mp3 = requests.get(mp3_url)
-                with open(mp3_file_path, 'wb') as f:
+                with open(mp3_file_path, "wb") as f:
                     f.write(response_mp3.content)
+
+            # write meta
+            meta_file_path = os.path.join(program_path, filename + ".json")
+            data = {
+                "id": id,
+                "title": title,
+                "description": node.get("description"),
+                "summary": node.get("summary"),
+                "duration": node.get("duration"),
+                "publishDate": node.get("publishDate"),
+                "programSet": {
+                    "id": programset_id,
+                    "title": node.get("programSet").get("title"),
+                    "path": node.get("programSet").get("path"),
+                },
+            }
+
+            with open(meta_file_path, "w") as f:
+                json.dump(data, f, indent=4)
         else:
             print("No programset_id found!", file=sys.stderr)
 
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(
-        description="ARD Audiothek downloader."
-    )
+    parser = argparse.ArgumentParser(description="ARD Audiothek downloader.")
     parser.add_argument(
-        "--url", "-u", type=str, default="", required=True, help="Insert audiothek url (e.g. https://www.ardaudiothek.de/sendung/2035-die-zukunft-beginnt-jetzt-scifi-mit-niklas-kolorz/12121989/)"
+        "--url",
+        "-u",
+        type=str,
+        default="",
+        required=True,
+        help="Insert audiothek url (e.g. https://www.ardaudiothek.de/sendung/2035-die-zukunft-beginnt-jetzt-scifi-mit-niklas-kolorz/12121989/)",
     )
     parser.add_argument(
         "--folder", "-f", type=str, default="./output", help="Folder to save all mp3s"
